The Makefile has been adjusted in order to allow compiling libImisIO, as well as some cleanup.
Namespaces are now used (namespace IOUtils) so that only 1 nodata is defined and used. A few new fields have been added to Meteodata (namely, hs and rswr). 

The demo, test program has been updated so that it works with the current version of the library (but ImisIO is not usable yet).