The Grid2DObjects have been modified by adding a latitude and a longitude field. These have to be provided alongside the other parameters, the goal being to consider that the east and north coordinates are convenience coordinates (for example for calculating distances), but that the reference coordinates are latitude/longitude.
The required changes have been implemented into BoschungIO and A3DIO. The latter now also computes latitude/longitude automatically out of the CH1903 coordinates (after checking if the lat/long are not already provided, like for meteo1d). 

The required formulas for implementing a local grid system have been put into IOUtils. These allow to compute the distance between two points, given their lat/long or the coordinate of a point given a starting lat/long, a distance and a bearing. This allows to define a local grid system, for example using the lower left corner of the dem as the origin.

The convertString for Date_IO is now used by meteoio_demo and createA3DFiles instead of a hand made equivalent. 

The Makefiles have been tweaked in order to remove unecessay dependencies (in test) and to add a missing compilation rule (main Makefile).