Added a hack allowing to read lat/long in the DEM as lower left coordinates.
The code for optional reading of wind directions has been activated in A3DIO. Some dead, commented code has been removed. 

The first version of the required code for fancier wind interpolation has been added. This highlights the need for a redesign of the structure of the spatial interpolations interface.