A new class has been created, DEMObject. It inherits from Grid2DObject but also contains DEM specific functionality: computation of slope, azimuth, normal vector's components and curvature.
A new method has been created for both DEMObject and Grid2DObject: the sub() method that allows extracting a subset of a given grid (usefull for distributed computing).

The Grid2DObject class has been tweaked in order to make it rely on IOUtils::nodata instead of its own nodata. The required changes have been made in A3DIO and libinterpol2D.

The Makefile.in and configure.ac have been adapted and slightly improved (still a long way to go...)