In order to make using the spatial interpolations easier to use, there is a need to be able to write to files objects of type Array2D. This has been added to IOInterface and implemented in the different plugins.
meteoio_demo has been modified accordingly and is now much simpler.

More fundamentally, we need to think about the way to handle in a nice and elegant way the different type accross the plugins (typically as seen for Write2DGrid): where should we handle what in order to let the user seemlessly write objects to the available mediums?