A parameter has been added to DEMObject::update in order to specify which algorithm to use. It can either take a member from a specifc enum or a string like "HICK" or "CORRIPIO".
The isSameGeolocalization method has been added to Grid2DObject in order to allow comparing the geolocalization attributes of two grids. It does not compare the xllcorner/yllcorner parameters since these might have been tweaked for convenience. The code checking the match between lat/long and xll/yll has been commented out since this prevents the correct operations when using a local grid. This method (checkCoordinates) will have to be completely redesigned.

The ISWR and LWR methods have been uncommented for the Meteo2DInterpolator. 

The (very dirty) hack that would have interpreted the grid coordinates differently depending on their value (as lat/long or swiss coordinates) in ARCIO and GRASSIO has been removed (this is now replaced by the much cleaner projection class). If the COORDPARAM key is missing in the io.ini file, it is quietly ignored instead of throwing an exception.

The meteoio_demo program has been updated to reflect these changes