Tow methods for getting the min and the max value out of a matrix have been added to Array, Array2D and Array3D. The updateAllMinMax of DEMBoject has been updated to rely on that instead of doing it by itself.
The doxygen comments of DEMObject have been fixed and the config.dox (doxygen configuration file) has been fixed (it was not processing .cc files). 

A new (private) methods has been added to DEMObject: safeGet(i,j) that returns a grid value for any (i,j) (so nodata if outside of the grid). This could be the right strategy for correctly computing slope/curvature on the dem even in the case of "holes" in the dem... (to be continued...)