An error message that was tagged [I] has been retagged as [E] so the user might be a little more scared (in A3DIO.cc)
A new constructor has been added to Date_IO (copy constructor).

The necessary changes have been done to the autoconf files so that an new option has come: --with-proj4 (turned off by default). But the user of the library still has to add -lproj to his compiler command line (after the -lmeteoio).