The buffering system shows itself improved - a fine tuning of the performance is possible through the following tags: BUFFERSTRATEGY = [always | newinterval] # default: always BUFFERPERIOD = 1 2
The BUFFERPERIOD tag configures the amount of days before and after a requested date the BufferedIOHandler will try to buffer. The BUFFERSTRATEGY determines how the buffer reacts if it already tried to buffer a certain interval and another date out of that interval is requested: "always" means a rebuffering is always attempted, newinterval means that rebuffering is not attempted if a requested date falls into a previously extracted interval.