Two minor potential bugs have been fixed in SNIO.
The A3DIO plugin now has a writeMeteo() method. It writes all timesteps, even if truncating the timestamp to the hour. It also assumes that all stations have synchronized timesteps (ie: reading the timesteps for the first station is enough). It can also not handle metadata that change over time and it does not split 2D files per years...