Fixed some doxygen comments in SNIO.
Some commented out code has been removed. The check for the projection of the stations compared to the projection used for the DEM has been re-activated (it MUST be active!). The first implementation of a custom grid "interpolation" method has been done (issue 50) but can not be used: we need access to an IOHandler in order to properly read grids provided by the user. Otherwise, the rest of the code for it has been written and tested.