Several bugs have been fixed in Coords: the setEPSG was not working, the altitude was not displayed with the << operator and the formatting of decimal_to_dms was broken in its handling of seconds...
The documentation for the resampling has been expanded and several other modules' documentation have been updated so their structure is built along the same lines. 

An error message in IOHandler (when failing to load a plugin) was not clear and has been changed.

Several bugs in SMEIO have been fixed (improper column names mapping, improper handling of easting/northing, unclear error message).

The example data set has been tagged as GMT+1 and the meteo_reading and 2D_interpolations modified accordingly (so the argument read on the command line is assumed to be in GMT+1 too)