The DEMObject subset constructor was broken, it has been fixed. It involved creating a subset method in the arrays classes.
The Coord class now supports using the given UTM zone when converting from lat/long to UTM. If the computed zone does not match the one requested in the io.ini, it prints a warning but uses the user-supplied zone.

A first implementation of readStationData for SMETIO has been done. 

Some documentation has been added for plugin development