A severe bug has been found in the spatial interpolations: the algorithm arguments were always the ones from the last evaluated algorithm (therefore not necessarily matching the chosen algorithm).
The MAGNUSS spatial interpolation algorithm now takes as an optional argument the name of the algorithm to use to initialize the grid. 

It is now possible to specify a lapse rate to be used for the CST_LAPSE algorithm (optional argument). If the "soft" argument is added, then a lapse rate is calculated from the data, and if it fails the user provided lapse rate is used instead. Without "soft", only the user provided lapse rate is used. Without any option, only the lapse rate calculated from the data itself is used.