The array getSum() was not the most appropriate method... It has been replaced by a getCount(), which means that to really get the sum, one has to do getMean()*getCount().
A new Coords::isNodata() method has been implemented as well as the possibility for two Coords objects to be equal if both are nodata. This is now used in SMETIO so that if the position of a station becomes nodata, it won't force its position information into the data section.

Some comments have been cleaned up in IMISIO, better exceptions thrown in SNIO and better backtrace formatting in IOExceptions (this being only for Linux, a dark grey text color has been specified to more or less make the backtrace less intrusive and scary to the user).

More Visual C++ tweaks in CMakeLists