The POPC version has been heavily transformed. Since we would need IOManager has a popc object and this is quite hard (we need a backreference in Meteo2DInterpolator), all parclasses have been removed. Instead, the user will have the responsability to packaged IOManager in his own parclass if needed. This makes the code simpler and skips difficult issues...
A new fill method has been added to the Array2D and Array3D templates that does the reverse of subset (ie: reassemble arrays).

Better error messages have been put in SNIO.

A new Date setter that takes a Date object has been created.

This version does NOT correctly compile with the autotools (it leads to a segfault when running code using MeteoIO), but the cmake compiled library works fine.