Two new methods have been implemented in IOUtils: bearing(string) that converts a string like "NW" into a numeric bearing and isNumeric(string) that returns true if the given string represents a number.
GSNIO has been modified so that the slope bearing are supported, both numeric and as strings (ie: "SSW" for example). The GSN field "NAME" is also mapped to StationData::name.

Finally, an option has been added to cmake, so the user can chose to compile as static library (now false by default).

And as usual, I forgot to update meteoio.lsm when doing the release...