Another fix/enhancement for POPC: the libraries now ends with "popc" in the popc version, as the autotools build was doing.
Windchill and heat index have been implemented in Atmosphere.

Some cleanup in the filters, some code moved to ProcessingBlock, so that elements that are not filters can inherit ProcessingBlock and get all that it needs. A processing element has been implemented to multiply by a given constant, another one to add a given offset and another one to correct precipitation for undercatch.

Some invalid doxygen comments have been fixed in MathOptim.