Two new methods have been implemented as static of Coords, to compute the ground length of one degree of latitude or longitude at any given latitude.
A check for 0x0 sized grids has been added to PNGIO. The ability to specify the number of colors (ie: discretization) in the io.ini has been added.

A breakthrough occured with GRIBIO, that starts to be able to do some things. This is still not usable, but definitely moving forward!