improvements for the Legend class: if the generated legend text is too long, an alternate format is used. Moreover, values close enough to zero are rounded to a true zero.
First working version of GRIBIO!! The file naming scheme might still be non-standard, but it can open a file, index it and access the various fields through the index (ie: it is very fast). When a given field is not available, some backup strategies are tried (like extracting TA and dew point if RH can not be directly read). Specifying a file name as argument does not work however, since such a file is a collection of grids (-> which grid to output?). One option might be to accept it and return the first grid (relevant if the file contains only one grid). A readDEM will be implemented later, as well as the ability to read virtual stations data.