The wind velocity for the undercatch correction was taken as is. It is now corrected to 2m using a log profile with a constant roughness length (0.03).
GRIBIO is now able to detect duplicate grid points when associating grid points and user virtual stations, and removes the duplicates (ie: the first station is kept, the other ones leading to the same grid point deleted). This means that before using it, the user should try a simple read_meteo to make sure that his points lead to unique virtual stations.