A wrong commit on doc/examples/io.ini has been reverted. A new undercatch filter has been implemented, following (Hamon, 1973). I would not recommend using it for the moment, though.
The issue 222 was actually more complex than it seemed: the behavior of an uncaught exception is actually not specified by the standard, so there is nothing that could be doneon MeteoIO's side: instead, each application is invited to properly wrap its calls in try/catch blocks, or wrap its whole main in try/catch. The code examples and examples have been modified to show several such possibilities.