The DataGenerators handling of missing meteo fields has been changed: if a generator is defined for a field that is not found in the vecMeteo, it will silently be skipped.
Since some grumpy people seem not to appreciate the graphical message box for exception messages, the ability to turn them off when compiling has been added to cmake. The code for generating such message boxes on Linux purely relying on libX11 is also committed (the boxes are ugly but they don't require more than libX11 and its headers).

We still have to make sure that enabling it by default (that would be the goal) does not make it more difficult to compile meteoio...