Removing the passing of ostrinstream objects by reference, since this forces us to reset the stringstream back to its original state and this lead to some compilation issues on some macs. Now, the interpolation messages are passed as normal strings.
Preparing the way for using IDW with an exponent on the distance. This is currently hardcoded to 1 (bypassing the call to "fast_pow" if it is 1 in order to have no performance impact in this case). We still have to figure out how the user could provide this value as an argument... In terms of performance, it is around 3% slower when calling fast_pow(x, 1.) compared to skipping this call. On some other tests, it was more on the order of 8%.