The POT_RADIATION generator has been renamed as ALLSKY_SW for consistency and its documentation updated. The ALLSKY_LW generator's documentation has also been updated.
A getHash() method has been added to the StationData. It is currently quite unsophisticated but should do the job. A better error message has been found for the "frac" lapse rates that can lead to a zero lapse rate (this is the case when there is no precipitation, for example).