A call to readDirectory now uses the pattern anywhere in the name, both for Posix and Windows.
The ALPUG plugin now uses a better way of validating the provided station_ids, by checking if at least one file matching the station ID can be found in the METEOPATH. If not, most probably the station ID is wrong. If yes, then some years are simply missing (and no error message or warning is generated). The silent skip of duplicates is now commented out in order to shame the owners of messed up data.