A bug has been fixed in MeteoGrids (a definition was missing).
A bug has been fixed in the NetCDF plugin that prevented reading some DEMs. The ability to remap parameters (between NetCDF and MeteoGrids) in the ini file has been added in order to add flexibility. The support for "years" and "months" time units has been implemented in a way that is compatible with the standard NetCDF tools.