A bug has been fixed in CosmoXMLIO: it was reading all the data with a time-shift (ie always reading the file *after* the timestep that it should have read).
A few documentation glitches have been fixed in IOInterface and Date and an invalid default value for the time zones fixed in SMETIO (instead of taking 0 as intended, it was always nodata). The missing files for the new OSHD plugin are now committed, although the plugin is NOT READY to be used yet.