Fixed a possible segfault.
There was an inconsistency that could lead to segfault in the handling of large
clocks of comments at the beginning of a CSV file. Depending how the number of
header lines is understood (is it the whole number of comment lines + "true"
header or only the pure header part?), it could have been that the date and time
parsing class would not be initialized. Now the minimum number of lines to read
as sneak peek into the file is dynamically adjusted and a check on the date_cols
object is performed to be sure that it has been initialized.

However, the more fundamental question whether a large clock of comments at the
beginning of  the CSV file belongs to the headers or not remains open and must
be clarified in the future...