Starting a new filter to measure a time shift between two correlated meteo parameters and be able to correct it
Basically, a Pearson's coefficient is optimized over a small sliding window in
order to extract a variable time offset between two meteo parameters that should
exhibit a high correlation coefficient. Then in a second run it will be possible
to apply a correction to bring the two meteo parameters back in sync.