Added signal trapping
The Posix signal SIGTERM is now trapped by meteoio_timeseries in order to
trigger the printing of a stack trace. This means that something like "kill
{meteoio_timeseries process}" stops the process and prints a stack trace (making
it easier to understand where a seemingly stuck process was). This will be used
by operational systems in order to kill processes that might have been running
for too long and still getting some useful debug information (we've had multiple
long running processes in the last few days, I suspect because of a unreachable
mount, this should help in validating or not this hypothesis on the running
system if this happens again).