Modernizing the support of libProj
Now relying on a FindProj.cmake macro in order to find libproj (there was no way
to use the files provided by the libproj project, therefore we have our own file
for now), renamed all instances of PROJ4 into PROJ (as this is the new name of
the project from version 7 onward) including in the ini files (so the ini files
MUST be migrated). This should make a smoother transition to the newer version
of libproj at a latter time (since many distributions still ship proj version 4
or 5, we might have to keep using the deprecated api for a long while).