Bug fix
In a test case provided by Charles, aggregating as the maximum was wrong
when the aggregation period was a multiple of the sampling rate.

As IOUtils::seek() returns either an exact match OR the next element
greater than the provided date, the get_window_specs() method was
inaccurately assuming that it was always the next element and always
substracting 1 to be sure to contain the start point of interest. Now
the case of exact match is properly handled and skips the substracting 1
in this case.