Suppress misleading 2D interpolation message
It does not look like the best spot to print this message... but for now
as long as all 2D interpolations return an info string it should be ok,
and it will not be printed if the data is read directly from disk
without interpolation.
In any case, it's cosmetics only.