Implemented support for cloud cover in All Sky ILWR generators.
This seems to work but should definitely be validated at a location with ILWR
measurements. In the MeteoBlue plugin, the radiation fields now rely on their
"instant" variant that matches what we want so this can be compared to our
internally computed potential radiation (the backwards variants had a time
shift, so they are probably averages over the past period or something similar).
Also some minor optimizations in the MeteoBlue plugin (including requesting
compressed data transfers if possible and switching to https).