Improving the behavior and performance of the ILWR parametrizations.
Several problems have been found when benchmarking the ILWR parametrizations at
the *WFJ station over the 2017-10-01 - 2021-10-01 period (with hourly data):

1. It happens surprisingly more often  than thought that ISWR_meas > ISWR_pot.
This was handled by returning nodata for the cloudiness, in order to temporally
interpolate (if possible) the cloudiness from neighboring points. This was
overly conservative and lead to degraded performance (for example when happening
before the night). Now in such a case the clearness index is reset to 1 so the
cloudiness gets set to a very low value or most often, zero.

2. There are even more problems around dawn/dusk than envisioned... So instead
of only computing the cloudiness when the radiation levels are above 5 W/m2, the
threshold is now 40 W/m2 (so if a sensor has some trouble with low solar
elevation it will not generate a very wrong cloudiness that would be used as a
basis for temporal interpolation over the night, thus generating a large number
of quite wrong data points). This threshold is now a public member of the
Atmosphere class in order to be broadly reused (for consistency).

3. Finally, it seems that the choice of cloudiness parametrization that
originally seemed a no-brainer is much more tricky... Officially, the clearness
index to cloudiness parametrization it tightly coupled with the cloudiness to
ILWR parametrization. But for our stations, even if not that far from one
another and is seemingly similar conditions, swapping cloudiness
parametrizations independently of the ILWR parametrization seem to make sense
for some cases. So this possibility has been added to the ILWR all sky
parametrization (by default, the cloudiness parametrization is the one that
belongs to the ILWR parametrization that has been chosen).