This is a backport of an older version of the cloudiness computation.
Fabiano had some trouble with his stations, where the parametrized ILWR was
quite different from what it used to be a while ago. Actually, it turns out that
this would only be visible when explicitly configuring a data generator for All
Sky ILWR and when the measurements had some problems (lots of abnormally low
ISWR values, probably due to snow covering the sensor). These kind of issues
have not been seen at WFJ.

So for now, the ISWR threshold has been reverted to 5 W/m2 and the whole
cloudiness computation to this older version. This means that the results at
locations such as WFJ are degraded compared to some of the last commits.