Another version of the cloudiness parametrization.
Now the criteria for daytime versus night or dawn/dusk is based on the solar
elevation. This works much better since now measurements problems such as snow
on a sensor lead to high cloudiness values (which is most probably the case,
otherwise the snow on the sensor would have melted) while keeping the improved
behavior at dusk/dawn that leads to more realistic values when interpolated over
the night. The performance plots will be redone with this new version.