Code clean up and new options
It is now possible to use the NOW time specification both for the start and the
end date options. It is also now possible to only provide one of these as well
as a duration (in days), for example setting an end date of 2022-01-31 with 30
days duration would compute a start date of 2022-01-01.