Bringing the horizons to the ILWR all sky generators.
Better handling of the TauCLDGenerator arguments vs AllSkyLWGenerator arguments:
the TYPE argument of TauCLDGenerator has been renamed so it does not conflict
with the TYPE argument of AllSkyLWGenerator.

A new file format for providing horizons has been designed and is used both by
ProcShade and TauCLDGenerator / AllSkyLWGenerator. The horizon can now be
computed from the dem, read from such as horizon file and written back to a
file.

Still to be done: make sure the documentation is up to date everywhere, update
inishell, check that the computed horizons make sense, check the impact on the
generated ILWR.