Allow refering to variables in an imported ini file
Doing the "local value has priority" in a different way has allowed to define a
key as a reference to a key that is imported or even defined with a call to
addKey() (that has to be performed before parsing the file that refers to it).
This gives more flexibility when writing ini files that must be relocatable
based on environment variables of a additional key such as "CWD", for example to
be used in sandboxed environment.