Fixed the writing of varying number of parameters in SMET
Handling varying number of parameters over a large time range has always
been tricky, as it was necessary to figure out the exact list of
parameters before allocating the memory as well as before writing the
data. The memory allocation part was done in a previous commit, but the
SMET plugin did not manage to reliably write such data.

This is now fixed, knowing that the parameters are now writen in
alphabetic order.