Fixed a very serious bug in attempting to handle indexed keys
Keys such as "STATION2", "STATION15" are considered indexed keys. A previous
commit attempted to properly sort such keys so key 15 comes *after* key 2. This
was done by extracting the index and sorting the keys based on the numerical
value of the index. But depending on the matching pattern that has been given,
we might attempt to sort keys such as "TA::GENERATOR1", "RH::GENERATOR1"... In
this case, both keys were considered to have the same index "1" and thus the
last one would overwrite the previous one. Now, this is properly handled by also
storing the "key root", here "TA::GENERATOR" and "RH::GENERATOR" so different
roots with the same index will not overwrite each other any more (while still
sorting keys that share the same root in numerical order).