A new otpion has been added to the NetCDF plugin: NC_ALLOW_MISSING_COORDS
This comes handy for non-conforming files although a data creator must be
configured in the [dataEditing] section otherwise I expect something to blow up
later on (the coordinates are assumed to be available).

Small trick in libncpp::readVariableMetadata() in order to look for time units
in the "description" field if nothing has been found in the "units" field.