The SYNTH plugin has gained much improved configurability and abilities. A new method has been added to Config
It is now possible to define how to generate meteo parameters in the
SYNTH plugin, per station and per parameter. Currently two methods are
implemented (constant and step functions), more will come.

A new getValuesRegex() method has been implemented in Config. It might
replace the getValues() method as it is more powerful but this will need
to be carefully validated...