All calls to fabs() have been replaced by calls to std::abs() for consistency
Adrien pointed out that the behavior of the different calls to abs (the c++
std::abs() and the C abs()) are not always equivalent (the c version might
return an int when a double was expected, depending on the compiler). So first,
calls to C's abs() have been replaced by std::abs() and now even calls to fabs()
(which are not ambiguous) have been replaced by std::abs() for consistency. See
https://stackoverflow.com/questions/21392627/abs-vs-stdabs-what-does-the-reference-say
for more on this topic.